requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic>=2.5.3
openai==1.93.0
PyMuPDF==1.24.1
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
streamlit>=1.36.0
tavily-python>=0.3.7
azure-ai-documentintelligence>=1.0.0b4
streamlit-drawable-canvas>=0.9.3
Pillow>=9.0.0
pandas>=1.5.0